backend/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
python-dotenv==1.1.1
requests==2.32.3
google-generativeai==0.3.2
supabase==2.23.0
python-multipart==0.0.9
Pillow==10.4.0
groq==0.8.0
